refactor: refactor how we read email cookies
Sometimes we need to pass the users email from one page to the next page
so that the email can be used.
In gateway there are actually two places where this email can be set.
a) the GU_email cookie set in gateway
b) the PLAY_SESSION_2 cookie set by scala play (identity-frontend)

This commit moves the functionality to handle email cookies into it's
own file, and creates a new method for reading email cookies, which
first tries the GU_email cookie followed by the PLAY_SESSION_2 cookie

We need to check both cookies as for now we still have routes that go
between the old identity-frontend platform and gateway