fix: update ordering of password eye icons
Changes the ordering of the password eye icons on the password input
component, to the following:

1. When password is not visible, show the eye icon to make it clear to
the user that clicking the icon will show the password.

2. When password is visible, show the eye icon with strike to make it
clear to the user that clicking the icon will hide the password.

This was achieved by simply making `isOpen={passwordVisible}`. This also
makes it more semantically readable, as "open" is associated with
"visibility".

These statements have also been added as comments to the code.

I've also wrapped the icons in an if-else statement to make it easier to
read what it happening.