feat: add ability to change password input component label text
As part of the design review for MVP0 in the migration project, we
want to update the text on the password input form from "New password"
to "Password".

However this same input form is used in multiple places in the project,
such as the change/reset password flow, where we want to keep the old
text.

To achieve this, this commit adds a `labelText` prop to the
`PasswordForm` component which can be used to customise the input label
text.

It then updates all instances of the `PasswordForm` component to add the
`labelText` prop in.