fix: fix ophan referral using viewId
When investigating why the `referrer_page_view_id` was not being
populated in the data late, we found that the way we were sending this
parameter to ophan was incorrect. Instead of sending just the `viewId`
as a parameter using ophan's `record` method, this should actually be
called `refViewId` according to the Ophan source files:

https://github.com/guardian/ophan/blob/70b658e785c490c411670bbd3c7fde62ae0224fc/the-slab/app/extractors/ReferrerExtractor.scala#L129

We also realised that we needed to pass a second parameter to Ophan too.
This is the `ref` parameter, which is the url of the referrer page.

https://github.com/guardian/ophan/blob/70b658e785c490c411670bbd3c7fde62ae0224fc/the-slab/app/extractors/ReferrerExtractor.scala#L171

This commit renames `viewId` to `refViewId` and adds the `ref`
parameter. We then make sure that we're sending these parameters as
expected by Ophan.

This commit also adds a method to validate the `ref` url parameter to
make sure that it's from a Guardian domain. Which is why
`validateReturnUrl.ts` was renamed to `validateUrl.ts` as this now
exports multiple method.