fix: ophan `ref` tracking on ophan init
Despite the efforts in https://github.com/guardian/gateway/pull/906 we
noticed that the referral information was still not pulling through in
Ophan and the data lake.

After some research we discovered that the `ref` parameter (the url of
the referrer page), is only able to be recorded on the init method of
the `ophan-tracker-js` library.

https://github.com/guardian/ophan/blob/fb36f3dab919abb482b453bdf6b6fbd5e60e7a44/tracker-js/assets/coffee/ophan/core.coffee#L18

However there was no way of modifying the values directly of what was
sent in this init method.

We then discovered that using `localStorage` it is possible to override
the parameters sent to ophan in the init method.

https://github.com/guardian/ophan/blob/fb36f3dab919abb482b453bdf6b6fbd5e60e7a44/tracker-js/assets/coffee/ophan/core.coffee#L66

This means that this commit removes the previous way that we were
sending the `ref` abd `refViewId` parameter. Now before ophan is
imported and initialised, we get the `ref` and the `refViewId`
parameter and add them to localStorage key `ophan_follow` if they exist,
which will be picked up by ophan, and then sent across in the when ophan
is imported.

This is done in the same way that the `support` platform sends this same
information.

https://github.com/guardian/support-frontend/blob/3a1ff437be200f2e37013b519ed037566be67b9c/support-frontend/assets/helpers/tracking/ophan.js#L116-L126