feat: allow a user to register a guest account in identity through Gateway.
Connect  to the /guest endpoint in IDAPI, forwarding ref, refViewId and returnUrl as part of the guest account registration. Implement a useRecaptcha hook to integrate with the Google Recaptcha service. Implement changes to match new designs, including a modification to the Divider component to allow it to fit to the width of the form.