feat: fix responsive design to match figma breakpoint definitions
feat: Add source kitchen and use where possible

The source development kitchen has recently been released with the aim
of speeding up component development.

https://github.com/guardian/source/tree/main/packages/%40guardian/source-react-components-development-kitchen

This commit adds the library as a dependency, and replaces existing
components to those that have been moved to the kitchen where possible.

fix: remove Divider changes, to be added back once added to the kitchen