feat(nav): update nav to match grid using flexbox
This upadtes the Nav (navbar) component to match alignment against
the grid layout. The specific breakpoints, size, padding, margins
etc. have been chosen as that is what is on dotcom-rendering at the
time of this commit.

We use flexbox instead of grid, as it provides more flexibility in
1D alignment. And because of the fact that the tabs are aligned to
the middle of the grid gutter/column-gap rather than the start of
the column.