refactor(emails): refactor emails to pass in profile url
This is because the remote storybook wasn't generating the stories for this email as the profileUrl was unable to be derived due to missing environment variables.

The template of the emails should not rely on calling external methods, we should pass in any values as a prop so that we can render these independently.