fix: add display: flex to ConsentsBlueBackground
On short pages using the ConsentsLayout and the ConsentsBlueBackground component, the border lines
would not strech to the bottom of the page.

To fix this we allow the ConsentsBlueBackground to act as a flex container, which means that children become flex items allowing them to grow to fill the remaining space.