refactor: GU_email to GU_GATEWAY_STATE cookie, that is encrypted
We want to be able to securely store state data on the client, that the
client will not be able to read or modify.

To facilitate this we created the GU_GATEWAY_STATE cookie, which will be
a httpOnly, secure, signed, and session cookie. However although this is
a highly secure cookie and unmodifiable cookie, it's still possible to
read the payload on the client.

So we encrypt the cookie using AES before it is set, and then decrypt it
when we want to read it.

We already has the GU_email cookie that was used for a similar purpose,
but we would like to add more secret data to the client. So the
functionality of the GU_email cookie was moved to this GU_GATEWAY_STATE
cookie.