feat (footer): update footer to match grid using flexbox
This updates the Footer component to match alignment against the grid layout.

We're using flexbox instead of grid, as it provides more flexibility in 1D alignment, as some of the footer tabs are aligned to the middle of the grid gutter/column-gap rather than the start of the column.
Using css-grid it's not (easily) possible to position items at the middle of the grid gap.