feat: refactor `POST /register` and setup email sending to existing users without password
Refactor the `POST /register` endpoint so that before it calls the guest endpoint to create the user, it instead calls the IDAPI `/user/type/:email`, to check if the user is either:
a) `new` - user not in Identity DB
b) `current` - existing user with password
c) `guest` - existing user without password

Based on the result of this `userType` we determine which action to take, i.e create a new user, or send one of the account exists emails depending of the user has a password or not.

This stops us relying on the error message of the `guest` endpoint so lets us clean up the error handling.

This commit also adds in the `/user/send-account-exists-without-password-email` IDAPI endpoint, so we can send the create password email to existing users who don't have a password.