feat: "ChangePassword" component changes
Update the "ChangePassword" component to use the latest copy, and add the ability to customise the "headerText" and "Confirm new password".

As part of the MVP1 flows, there are two seperate routes which change the users password.

a) The existing "reset/change password" route for users going through the reset password flow
b) A "set password" route for users without an existing password, e.g guest users

Since these routes are mainly identitcal, it would be useful to reuse the existing ChangePassword component.