feat: make EmailInput component and test, add upgrade to kitchen v0.0.20.
This component is to be reused across pages where an email input is
required that does validation checks on blur. Currently our requirements
stipulate that we should handle cases where the input is empty or has an
invalid email. This is handled by some internal state. The component is
rendered and its behaviour is unit tested using react-testing-library and
the jest-dom extension.