fix: return url timeout issue
Redirecting to the /about page caused this test to intermittently time out, it's very likely to be the same issue as #1033 where the fix was to redirect to a simple CAPI JSON response for the page. This takes a similar approach, rendering the CAPI JSON output for the about page.