Email Input component (#1061)
* feat: make EmailInput component and test, add upgrade to kitchen v0.0.20.

This component is to be reused across pages where an email input is
required that does validation checks on blur. Currently our requirements
stipulate that we should handle cases where the input is empty or has an
invalid email. This is handled by some internal state. The component is
rendered and its behaviour is unit tested using react-testing-library and
the jest-dom extension.

* feat: cover both possible error states in test for valid email

* fix: use the HTML 'validity' object on the email input.

We use it to determine whether the email is
valid or not. Also suppress the default
browser popup using preventDefault on the
'invalid' event. Expand test to include more
thorough check and some general refactors.

* fix: add customError invalid check after review