refactor(error): Differentiate between Api/Idapi errors and runtime errors
We don't want to propegate runtime javascript error messages to the
user, as was potentially ocurring when destructing the `error` object
in many of our routes, and instead provide a generic message and
status back to the user.

To differentiate Api/Idapi errors from runtime javascript error, we
create a new `ApiError` class and `IdapiError` (which extends
`ApiError`) class, and replaces the existing `RequestError` interface.

Then when we want to check the `message` and `status` from the
error we can do `error instanceOf ApiError` to determine if it's an
`ApiError` or `IdapiError`. If it's neither then we instantiate
an empty `ApiError` class, which returns a generic message.

We'll still be able to view runtime errors as we log out the whole
error object when an error is caught.