feat: Add `ExternalLink` and `ExternalLinkButton` components
We want to apply `noreferrer` and `noopener` on external links for
security reasons.

We could easily apply these to the `rel` attribute on the source
`Link`/`LinkButton` component, but to be more expicit that the link
is pointing externally, we create a component that extends the link
component with the attribute already applied.