feat: make `QueryParams` available on `ClientState`
We seem to have trouble when reading query parameters on the
client side of gateway. Often passing them in as extra properties
on page data, or attempting to read them using `location.search`/
`useLocation` etc.

By adding them as a property on the `ClientState` it would make it
easier to read and use the query parameter properties.