refactor(routes): use typed array for client routes
With the update to react-router 6 it's no longer possible to use
a custom route component which we were using to type the path.

https://reactrouter.com/docs/en/v6/upgrading/v5#refactor-custom-routes

To get around this limitation we make an array which takes the
typed path and element, and then map over the array creating
a Route component from react-router avoiding the need for a custom
component, but still keeping type safety.