feat(tracking): add page load tracking to "ResetPassword" component
By sending an ophan interaction event on a page load/render we can track
the different variations of a page if they have the same url.
For example, a link expired page is indistiguishable from the url as a
valid password reset page. By tracking which page loads, we can measure
how many users see each variation of the page.