feat(tracking): add server side ophan tracking
Since there are some events which cannot be tracked using the ophan
client library, such as determining which email was sent to the user, we
have to send these from the server side.

The ophan library only supports browsers, so we have to manually create
a method that calls ophan in the same way. We call the ophan endpoint
with the `viewId`, `browserId` (`bwid`), at the minimum, and provide the
`consentUUID` value if it exists (since this is also being sent from the
browser on every request to ophan if available).

We get this values from the client, the `viewId` (`pageViewId`) is already
available in the tracking through the `refViewId` parameter, so we can
associate this event with a page view. The `browserId` (`bwid`) and `consentUUID` are
set as cookies on the *.theguardian.com domain, so we read these values
when they are sent to the server. To make these easier to work with we
created an OphanConfig interface on the RequestState.

We can then use `fetch` (`node-fetch` library) to make this request to ophan, using
`'timeout-signal'` to timeout the request if it takes too long (>250ms).
The events that can be sent are the same as what's available on the
client through the `OphanEvent` interface.