fix(safari10/11): use only legacy bundle to fix js issues
Safari 10/11 both support the `module` tag, but has bugs in the
implementation of some modern features, causing javascript issues and
making anything that relied on these features unusable.

In this case there was an issue with the use of promises and the
breached password check form.

We decided it was best to exclude the modern bundle from safari 10 and
11 and only rely on the legacy bundle. To do this we parse the user
agent and add the browser information on the request state. In the
renderer we check the browser version and include the relevant script
tags in the response.