fix: ie11 polyfill and transpilation issues
This commit fixes two issues with IE11:

1.
We were mixing polyfills provided by polyfill.io and corejs causing an
uncaught error to be thrown in IE11.
Ideally we want to move all polyfills to using polyfill.io rather than
corejs as we can target only features that the users browser does not
support rather than providing polyfills for all features using corejs.

As a temp migration measure we set the `type="module"` attribute on the
polyfill.io script tag so that it is not executed by legacy browsers
such as IE11.

We also add all supported features we transpile with to polyfill.io
namely: `es2015` through `es2021` and `fetch`. This includes polyfills
for `Object.fromEntries` and `Promise.prototype.finally`.

2.
Although we provide a polyfill for `Promise.finally`, our usage of it
causes a `Object doesn't support property or method 'finally'` only in
the case of the `PasswordForm.tsx`, after investigation it seems that
the transpilation is behaving in a strange way for this promise in
legacy browsers, where as in other cases `.finally` behaves as expected.

We solve this issue by refactoring the breached check to use async/await
avoiding the need for the `.finally` in this scenario.

We also add sentry logging for checking when the breached password check
fails for any reason.