feat: sign in session check using auth code flow
This commit modifies the sign in flow, so that when okta is enabled it
will first perform a check to see if an okta session exists by
performing the auth code flow.

If an okta session already exists then we complete the auth code flow
and redirect the user back to the returnUrl.

If a session doesn't exist okta return an `login_required` error in the
oauth callback, we intercept this error and use it to show the sign in
page. If there are any other errors that we log those oauth errors and
display a generic error message.

To fascilitate this we use the EncryptedState cookie to keep track if
the sign in session check is complete by setting the `signInRedirect`
flag when redirecting to the sign in page. If this flag is true then
render the sign in page.

We also make changes to the AuthorizationState by replacing the
`returnUrl` with the `queryParams` at the start of the request, this is
because we now track a lot more than just the returnUrl, for example for
ophan tracking with `ref` and `refViewId`.