refactor(crypto): pass encryption key into encrypt and decrypt
We want to be able to use the encrypt and decrypt methods in cypress for
running tests, however the use of `getConfiguration` meant that it was
reliant on code that cypress did not have access to.

This commit refactors the methods so you have to explicitly pass the
encryption key in when performing encryption and decryption. This allows
us to import this file outside of the server, for example cypress.

This makes these methods completly functional too.

We update the tests and usages accordingly.