feat: add explicit OktaSignIn metric, trackMetric return type void
This makes a slight change to the `trackMetric` method making sure to
explitly make the return type `void` as we want to not return anything
from this method, and make it fire and forget.

Although we haven't noticed this behaviour,
we were worried that if this promise failed
it would also be handled by a try-catch block higher up the call chain
as well as the promise catch.

By not returning anything and making it void, we can make sure that this
is more explicitly defined.