refactor: move client side ab testing logic outside react
This commit changes the way the logic for client side ab testing
behaves.

We first update the models so that the `ABTesting` type is
consistent between the `ClientState` and the `RequestState` and we add
on the `hasParticipations` property which is a boolean value determining
if this request/user/browser is currently in any tests.

On the client we move out the logic from react into pure js by switching
from the `@guardian/ab-react` to `@guardian/ab-core` and moving it into
it's own file and method, and attach the `ABTestApi` to
`window.guardian.abTestApi` which can be used elsewhere on the client.

If the user is in any tests, then we dynamically import this file/method
and initialise the AB test api. This reduces the overall browser bundle
slightly by only requiring `@guardian/ab-core` when needed.