feat: setup profile.thegulocal.com with Okta in DEV
We recently changed the Okta url to sit on the profile subdomain, while
the routing for this is easy to control in PROD and CODE thanks to vcl,
it's a bit harder to do locally/in the DEV environment as the OAuth/OIDC
tokens won't verify thanks to the `issuer` (`iss`) being different.

In https://github.com/guardian/identity-platform/pull/505 we set up the
local nginx configuration for profile.thegulocal.com to point okta to
the correct development environment. This is done through the use of the
`gu_okta_env` cookie which can be set to `CODE` or `DEV` depending on
which environment of okta you want to point to locally. This returns a
"X-GU-Okta-Env" header which we can use to control this `issuer` field.

So when `stage` is `DEV`, and the `X-GU-Okta-Env` exists, we set up a
new `OpenIdClient` with the correct `issuer` set so that the tokens we
get back from okta match the `issuer` checks.

This allows issuers to develop locally with Okta tokens and an Okta
session set locally too.