feat: use Zustand to share state between Islets
[Zustand](https://github.com/pmndrs/zustand) is a small state-management
library which is able to share state between different renderers i.e.
different islets/islands.

This commit is a barebones implementation of the library, just setting
up the store itself and the hook. Examples of usage can be seen in the
PR and documentation.

This is required as we want an action from one islet to update something
in another islet, this is not possible with React Context currently, and
we don't want to use a large state management solution (Zustand is
<1kb).