feat: add templateable section in AccidentalEmail
This email is sent by Okta when something has gone wrong, and can be
used for debugging. In the Okta email Terraform config, the HTML output
of this template is read and the {{CTA}} block is filled in with some
Okta $templateStrings (without these, Okta won't send the email). For
most users, these strings shouldn't be shown, so the {{CTA_DISPLAY}}
value is set to 'none' to hide this section. In some cases, we may want
to set {{CTA_DISPLAY}} to 'block' to show the section instead.