fix: Add hidden Okta variable to template
In identity-platform/okta/terraform/modules/okta/email_templates/
resetPassword.html, there is a display:none div with the text
${resetPasswordLink}, which Okta requires to be present in the email
when applying a Teraform config. This div was presumably added in by
hand to the MJML-generated HTML. To match the current functionality, I
have added a new prop, hiddenText, to the Footer email component, which
will generate a similar hidden div with whatever text string is passed
into it. In the ResetPassword template, this string contains
'${resetPasswordLink}'.