ci: reduce cypress tests in memory to possibly help flakiness
By reducing the number of tests in memory to 0, cypress does not keep
any previous snapshots of tests in it's memory, this is useful for CI as
we don't/can't inspect previous snapshots, and we just want to know the
result.

This has the added benefit of reducing memory usage, which might reduce
flakiness in cypress tests running in GitHub actions.