feat(okta): allow social users to reset their password
This commit makes a number of changes which allows a social user
(`user.credentials.provider.type === "SOCIAL"`) to be able to get a
reset password email to set a password.

By default Okta does not let you send a reset password email or generate
a password reset token for users with their provider set to "SOCIAL" or
"FEDERATION".

https://developer.okta.com/docs/reference/api/users/#provider-object

This means we have to get a user out of this provider to the "OKTA"
provider to be able to send them a password reset email. To do this we
have to set a placeholder password in a specific way.

Firstly we use this endpoint in the users lifecycle api
`'/api/v1/users/:id/lifecycle/reset_password`  which will force
the user into the "RECOVERY" state, change the provider to "OKTA", and
generates a password reset token (`recoveryToken`).

However users in the "RECOVERY" state will not be able to signin with
any method until the password is reset, so if a user doesn't complete
this flow then they will be locked out of their account. Instead of
sending this token to the user, we use this token to perform a password
reset flow on the server side and using a randomly generated, unknown,
placeholder password to get the user into the "ACTIVE" state.

Once they're in this "ACTIVE" state with the "OKTA" provider, we are
able to send them the normal password reset email which we use for all
other users.