feat(okta): social users email validated workaround
We're unable to set the `user.emailValidated` field in the Okta user profile for social users when they are created, but we are able to put them in the `GuardianUser-EmailValidated` group.

So this is a workaround to set the `emailValidated` field to true in the Okta user profile if the user is in the `GuardianUser-EmailValidated` group, based on custom claims we have set up on the `id_token`.

Specifically the `user_groups` and `email_validated` claim.

The `user_groups` claim returns an array of groups that the user is in, and the `email_validated` flag returns the status of the `user.emailValidated` profile field.

If the user is in the `GuardianUser-EmailValidated` group, but the `emailValidated` field is falsy, then we set the `emailValidated` field to `true` in the Okta user profile by manually updating the user object.