feat: Add /reauthenticate version of SignIn page
The /reauthenticate page is used as a place for users with old sessions
to be sent when they need to reauthenticate. The only difference between
/signin and /reauthenticate is that /reauthenticate doesn't have the
Register button visible in the navbar, so instead of creating a new
reauthenticate page, we just pass a displayRegisterTab flag to the
SignInPage component when we're serving /reauthenticate.