fix(okta): change password send email controller
Fixes issue with email sent page for change/reset password send controller not showing the "Resend button" or "Didn't receive an email text" for users who's accounts did not exist. This fixes an account enumeration bug.
To fascilitate this we just have to set the encrypted cookie in for this user type as well. Hence the new `setEncryptedCookieOkta` method to share this in the two places it's used.