feat: add ad free cookie if user has a product
Due to race conditions occuring on dotcom platforms (theguardian.com), it's possible for some users who have a valid product to still see adverts, on the first page view after signing in.
This is because the ad free cookie is set after the page loads and executes.

This PR attempts to set the ad free cookie on sign in or session refresh.

We do this by first calling members-data-api's `/user-attributes/me` endpoint to get the products that the user has.

We then check that the user has the `digitalPack` product, if they do then we set the ad free cookie (`GU_AF1`) on `*.theguardian.com`.

This code is based on code from a number of places used to implement this functionality.
- https://github.com/guardian/frontend/blob/main/static/src/javascripts/projects/common/modules/commercial/user-features.ts
- https://github.com/guardian/frontend/blob/main/static/src/javascripts/lib/manage-ad-free-cookie.ts
- https://github.com/guardian/members-data-api/blob/main/membership-attribute-service/app/models/Attributes.scala