feat: move emails to gateway
This moves the 'Complete Registration' and 'Reset Password'
emails to be sent from Gateway (they were previously sent by Okta
as a side-effect of Okta API calls). As a result, this also removes
the public routes we exposed which returned email template HTML
for consumption by the Okta Terraform scripts. The 'Verify' email
templates are renamed to 'CompleteRegistration' to better match
their use case.