feat: add webpack-bundle-analyzer
Add `webpack-bundle-analyzer` to our webpack config so we get visibility
over our browser bundle sizes and see where we can improve.

The bundle analyzer only runs on the `make build-analyze` or `yarn build:analyze` commands.

It will save the output to the `build` folder, and automatically open them up in a browser when generated.

Co-authored-by: Raphael Kabo <raphaelkabo@users.noreply.github.com>