refactor: remove `react-router`
ALl our routing was happening on the server side anyway, so we actually don't need to include `react-router` at all in the
project, thus reducing the bundle size drastically.

We replaced the router with `regexparam` functionality, which converts a path to a regex, and in the `routes.tex` file we
loop through the routes to see if any match, and then return the element to display if so, effectively what react router
was previously doing.

In other places we remove the `useLocation` and `useParams` hooks by replicating this functionality in different ways.

For `useLocation` we use the native browser APIs. And for `useParams` we instead pass down the `token` parameter into the
`ClientState` so we can read it on the client side.

We also update some cypress tests which broke as a result of the hook changes.

| | Modern | Legacy |
| - | - | - |
| `@remix-run/router/dist` | -37.04 kB (-12.37 kB gzipped) | -49.94 kB (-15.25 kB gzipped) |
| `react-router/dist` | -12.49 kB (-4.28 kB gzipped) | -16.28 kB (-5.11 kB gzipped) |
| `react-router-dom | -14.28 kB (-4.96 kB gzipped) | -20.94 kB (-6.07 kB gzipped) |
| Total | -63.81 kB (-21.58 kB gzipped) | -87.16 kB (-26.43 kB gzipped) |

Co-authored-by: Raphael Kabo <raphaelkabo@users.noreply.github.com>